Ajout de la configuration Spring

+++ WebService-Spring/src/main/java/fr/afcepf/atod26/webservice/spring/api/IServiceSpring.java
@@ -12,7 +12,7 @@ public interface IServiceSpring {
 			@WebParam(name = "deuxiemeParam") final int deuxieme);
 
 	@WebMethod(operationName = "diviserSpring")
-	int division(@WebParam(name = "premierParam") final int premier,
+	double division(@WebParam(name = "premierParam") final int premier,
 			@WebParam(name = "deuxiemeParam") final int deuxieme);
 
 }
